Default to ERROR_ON_UNDEFINED_SYMBOLS, even when SIDE_MODULE is set. NFC
This change removes the defaulting of ERROR_ON_UNDEFINED_SYMBOLS when
building a side module.

In order handle that I removed from REQUIRED_EXPORTS all of the symbols
that are not relevant when building a side module.

This change is basically NFC.  The main thing it fixes a lot of
incorrect "undefined symbol" errors when building a side module with
`ERROR_ON_UNDEFINED_SYMBOLS`.

Fixes: #17605